First commit
This is the CircuitPython code for the trinket (a test) and the QtPy
(production, if you will).

This was built into a little birdhouse as a gift. There's unlikely to be
a full build/description forthcoming. I'm committing this here so I
don't lose it, but I need to finish the build as a gift...